CLN: Move InitializeCase to own file
